+++ motion/transcode.c
@@ -56,7 +56,7 @@ int main(int argc, char* argv[]) {
 	av_log_set_level(loglevel);
 
 	int err = 0, ret = 0;
-	unsigned long components = 0;
+	uint8_t components = 0;
 	unsigned long widths[4], heights[4];
 	uint64_t nframes;
 	FFColorProperties color_props;
